part_01/requirements.txt
numpy==1.22.1
pandas==1.4.0
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
six==1.16.0
